Final suggestions
Re-read this. Looks great!

The only real issue is the presentation of the FAIR principles. The text from the wikipedia page makes it look like Findability,... etc ARE the principles and not what the principles ARE ABOUT, and then the wikipedia links for the terms are too general to make sense. So I tried to rewrite it slightly.

Rest are just small edit